% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_listdeps.R
\name{do_listdependents}
\alias{do_listdependents}
\title{list dependents of a given package}
\usage{
do_listdependents(pname = "ROC")
}
\arguments{
\item{pname}{character(1) name of package}
}
\description{
list dependents of a given package
}
\note{
This will push a revised version of this package to trigger the action.
}
